Withdraw Screne Finished
Created withdraw screen, and included buttons for different values.

Using userData attribute in the button allows for Withdraw.java to be smaller.

Withdraw_Amount in Withdraw.java is the current public value to retrieve amount the user wants to withdraw.

Had to remove the copyright as it kept moving, may include it later.

Updated README.md with approprate changes.