Added Confirmation Screen, Success Screen
Added both confirmation screen, and success screen as well as created an interface for the screen switcher so it is easier to switch scenes.

Planning in the future is to change the success screen with a dynamic screen that will change depending on actual success or failure.

Also updated README