GUI Mostly Finished
Added final scenes including Deposit, and Check Balance. Made them both functional.
Gave Confirmation screen a bit of backend to get the transaction type, as well as the amount. This also is now working for checking balance.

For the full backend, we will need to implement a way to check a file for the login information, as well as change the balance on that.

For the future, I will work on code optimization, and cleanup to make it look, and function better in prep for the full backend.

Updated README with new files / comments / spelling fixes.